+++ cmd/cli.go
@@ -16,6 +16,7 @@ import (
 	"github.com/statping/statping/utils"
 	"io/ioutil"
 	"os"
+	"path/filepath"
 	"strings"
 	"time"
 )
@@ -38,12 +39,15 @@ func systemctlCli(dir string, uninstall bool, port int64) error {
 	location := "/etc/systemd/system/statping.service"
 
 	if uninstall {
+		fmt.Println("systemctl stop statping")
 		if _, _, err := utils.Command("systemctl", "stop", "statping"); err != nil {
 			log.Errorln(err)
 		}
+		fmt.Println("systemctl disable statping")
 		if _, _, err := utils.Command("systemctl", "disable", "statping"); err != nil {
 			log.Errorln(err)
 		}
+		fmt.Println("Deleting systemctl: ", location)
 		if err := utils.DeleteFile(location); err != nil {
 			log.Errorln(err)
 		}
@@ -67,8 +71,8 @@ After=network-online.target
 [Service]
 Type=simple
 Restart=always
-Environment=STATPING_DIR=` + dir + `
-Environment=ALLOW_REPORTS=true
+Environment="STATPING_DIR=` + dir + `"
+Environment="ALLOW_REPORTS=true"
 ExecStart=` + binPath + ` --port=` + utils.ToString(port) + `
 WorkingDirectory=` + dir + `
 
@@ -78,26 +82,30 @@ WantedBy=multi-user.target"
 	fmt.Println("Saving systemctl service to: ", location)
 	fmt.Printf("Using directory %s for Statping data\n", dir)
 	fmt.Printf("Running on port %d\n", port)
+	fmt.Printf("\n\n%s\n\n", string(config))
 	if err := utils.SaveFile(location, config); err != nil {
 		return err
 	}
+	fmt.Println("systemctl daemon-reload")
 	if _, _, err := utils.Command("systemctl", "daemon-reload"); err != nil {
 		return err
 	}
+	fmt.Println("systemctl enable statping")
 	if _, _, err := utils.Command("systemctl", "enable", "statping.service"); err != nil {
 		return err
 	}
+	fmt.Println("systemctl start statping")
 	if _, _, err := utils.Command("systemctl", "start", "statping"); err != nil {
 		return err
 	}
 	fmt.Println("Statping was will auto start on reboots")
-	fmt.Println("systemctl service: /etc/systemd/system/statping.service")
+	fmt.Println("systemctl service: ", location)
 
 	return nil
 }
 
 func exportCli(args []string) error {
-	filename := fmt.Sprintf("%s/statping-%s.json", utils.Directory, time.Now().Format("01-02-2006-1504"))
+	filename := filepath.Join(utils.Directory, time.Now().Format("01-02-2006-1504")+".json")
 	if len(args) == 1 {
 		filename = fmt.Sprintf("%s/%s", utils.Directory, args)
 	}
@@ -135,7 +143,7 @@ func sassCli() error {
 	if err := source.Assets(); err != nil {
 		return err
 	}
-	if err := source.CompileSASS(source.DefaultScss...); err != nil {
+	if err := source.CompileSASS(); err != nil {
 		return err
 	}
 	return nil
@@ -190,6 +198,10 @@ func resetCli() error {
 
 func envCli() error {
 	fmt.Println("Statping Configuration")
+	fmt.Printf("Process ID:          %d\n", os.Getpid())
+	fmt.Printf("Running as user id:  %d\n", os.Getuid())
+	fmt.Printf("Running as group id: %d\n", os.Getgid())
+	fmt.Printf("Statping Directory:  %s\n", utils.Directory)
 	for k, v := range utils.Params.AllSettings() {
 		fmt.Printf("%s=%v\n", strings.ToUpper(k), v)
 	}