+++ frontend/src/components/Dashboard/ServiceInfo.vue
@@ -1,94 +1,56 @@
 <template>
-    <div class="card mb-4" :class="{'offline-card': !service.online}">
-        <div class="card-title px-4 pt-3">
-            <h4 v-observe-visibility="setVisible">
-                <router-link :to="serviceLink(service)">{{service.name}}</router-link>
+    <div class="dashboard_card card mb-4" :class="{'offline-card': !service.online}">
+        <div class="card-header pb-1">
+            <h6 v-observe-visibility="setVisible">
+                <router-link :to="serviceLink(service)" class="no-decoration">{{service.name}}</router-link>
                 <span class="badge float-right text-uppercase" :class="{'badge-success': service.online, 'badge-danger': !service.online}">
                     {{service.online ? $t('online') : $t('offline')}}
                 </span>
-            </h4>
+            </h6>
         </div>
 
-        <div class="card-body p-3 p-md-1 pt-md-1 pb-md-1">
-
-            <transition name="fade">
-            <div v-if="loaded" class="col-12 pb-2">
-
-                <div v-if="false" class="row mb-4 align-content-center">
-
-                    <div v-if="!service.online" class="col-3 text-left">
-                        <span css="text-danger font-5 font-weight-bold"></span>
-                        <span class="font-2 d-block">Current Downtime</span>
-                    </div>
-
-                    <div v-if="service.online" class="col-3 text-left">
-                        <span class="text-success font-5 font-weight-bold">
-                            {{service.online_24_hours.toString()}} %
-                        </span>
-                        <span class="font-2 d-block">Total Uptime</span>
-                    </div>
-
-                    <div v-if="service.online" class="col-3 text-left">
-                        <span class="text-success font-5 font-weight-bold">
-                            0
-                        </span>
-                        <span class="font-2 d-block">Downtime Today</span>
-                    </div>
-
-                    <div v-if="service.online" class="col-3 text-left">
-                        <span class="text-success font-5 font-weight-bold">
-                            {{(uptime.uptime / 10000).toFixed(0).toString()}}
-                        </span>
-                        <span class="font-2 d-block">Uptime Duration</span>
-                    </div>
-
-                    <div class="col-3 text-left">
-                        <span class="text-danger font-5 font-weight-bold">
-                            {{service.failures_24_hours}}
-                        </span>
-                        <span class="font-2 d-block">Failures last 24 hours</span>
-                    </div>
-
-                </div>
-
-                <div class="row">
-                    <div class="col-md-6 col-sm-12 mt-2 mt-md-0 mb-3">
-                        <ServiceSparkLine :title="set2_name" subtitle="Latency Last 24 Hours" :series="set2"/>
-                    </div>
-                    <div class="col-md-6 col-sm-12 mt-4 mt-md-0 mb-3">
-                        <ServiceSparkLine :title="set1_name" subtitle="Latency Last 7 Days" :series="set1"/>
-                    </div>
-                </div>
-
+        <div class="card-body">
+            <div v-if="loaded" class="row pl-2">
+              <div class="col-md-6 col-sm-12 pl-2 mt-2 mt-md-0 mb-3">
+                  <ServiceSparkLine :title="set2_name" subtitle="Latency Last 24 Hours" :series="set2"/>
+              </div>
+              <div class="col-md-6 col-sm-12 pl-0 mt-4 mt-md-0 mb-3">
+                  <ServiceSparkLine :title="set1_name" subtitle="Latency Last 7 Days" :series="set1"/>
+              </div>
+              <ServiceEvents :service="service"/>
             </div>
-            </transition>
+              <div v-else class="row mb-5">
+                <div class="col-12 col-md-6 text-center">
+                  <font-awesome-icon icon="circle-notch" class="text-dim" size="2x" spin/>
+                </div>
+                <div class="col-12 col-md-6 text-center text-dim">
+                  <font-awesome-icon icon="circle-notch" class="text-dim" size="2x" spin/>
+                </div>
+              </div>
         </div>
-
         <div class="card-footer">
-            <div class="row">
 
-                <div class="col-12 col-md-3 mb-2 mb-md-0">
-                    <router-link :to="{path: `/dashboard/service/${service.id}/incidents`, params: {id: service.id} }" class="btn btn-block btn-white text-capitalize incident">
-                        {{$tc('incident', 2)}}
-                    </router-link>
-                </div>
-                <div class="col-12 col-md-3 mb-2 mb-md-0">
-                    <router-link :to="{path: `/dashboard/service/${service.id}/checkins`, params: {id: service.id} }" class="btn btn-block btn-white text-capitalize checkins">
-                        {{$tc('checkin', 2)}}
-                    </router-link>
-                </div>
-                <div class="col-12 col-md-3 mb-2 mb-md-0">
-                    <router-link :to="{path: `/dashboard/service/${service.id}/failures`, params: {id: service.id} }" class="btn btn-block btn-white text-capitalize failures">
-                        {{$tc('failure', 2)}} <span class="badge badge-danger float-right mt-1">{{service.stats.failures}}</span>
-                    </router-link>
-                </div>
-                <div class="col-12 col-md-3 mb-2 mb-md-0 mt-0 mt-md-1">
-                    <span class="float-md-right">
-                        {{$t('uptime', [service.online_7_days])}}
-                    </span>
-                </div>
+          <div class="row">
+          <div class="col-5 pr-0">
+              <span class="small text-dim"> {{ hoverbtn }}</span>
+          </div>
 
+            <div class="col-7 pr-2 pl-0">
+              <div class="btn-group float-right">
+                  <button @click="$router.push({path: `/dashboard/service/${service.id}/incidents`, params: {id: service.id}})" @mouseleave="unsetHover" @mouseover="setHover('Incidents')" class="btn btn-sm btn-white incident">
+                    <font-awesome-icon icon="bullhorn"/>
+                  </button>
+                  <button @click="$router.push({path: `/dashboard/service/${service.id}/checkins`, params: {id: service.id}})" @mouseleave="unsetHover" @mouseover="setHover('Checkins')" class="btn btn-sm btn-white checkins">
+                    <font-awesome-icon icon="calendar-check"/>
+                  </button>
+                  <button @click="$router.push({path: `/dashboard/service/${service.id}/failures`, params: {id: service.id}})" @mouseleave="unsetHover" @mouseover="setHover('Failures')" class="btn btn-sm btn-white failures">
+                    <font-awesome-icon icon="exclamation-triangle"/> <span v-if="service.stats.failures !== 0" class="badge badge-danger ml-1">{{service.stats.failures}}</span>
+                  </button>
+              </div>
             </div>
+
+          </div>
+
         </div>
 
         <span v-for="(failure, index) in failures" v-bind:key="index" class="alert alert-light">
@@ -100,19 +62,20 @@
 </template>
 
 <script>
-  import Checkin from '../../forms/Checkin';
-  import FormIncident from '../../forms/Incident';
-  import FormMessage from '../../forms/Message';
-  import ServiceFailures from '../Service/ServiceFailures';
-  import ServiceSparkLine from "./ServiceSparkLine";
+  const Checkin = () => import(/* webpackChunkName: "dashboard" */ '../../forms/Checkin');
+  const FormMessage = () => import(/* webpackChunkName: "dashboard" */ '../../forms/Message');
+  const ServiceFailures = () => import(/* webpackChunkName: "dashboard" */ '../Service/ServiceFailures');
+  const ServiceSparkLine = () => import(/* webpackChunkName: "dashboard" */ "./ServiceSparkLine");
   import Api from "../../API";
+
+  const ServiceEvents = () => import(/* webpackChunkName: "dashboard" */ "@/components/Dashboard/ServiceEvents");
 
   export default {
       name: 'ServiceInfo',
       components: {
+        ServiceEvents,
           Checkin,
           ServiceFailures,
-          FormIncident,
           FormMessage,
           ServiceSparkLine
       },
@@ -125,6 +88,8 @@
       data() {
           return {
               uptime: null,
+              hovered: false,
+              hoverbtn: "",
               openTab: "",
               set1: [],
               set2: [],
@@ -138,7 +103,16 @@
     watch: {
 
     },
-      methods: {
+    mounted() {
+      this.unsetHover()
+    },
+    methods: {
+      setHover(name) {
+        this.hoverbtn = name
+      },
+        unsetHover() {
+          this.hoverbtn = this.$t('uptime', [this.service.online_7_days])
+        },
         async setVisible(isVisible, entry) {
           if (isVisible && !this.visible) {
             await this.loadInfo()
@@ -197,16 +171,3 @@
       }
   }
 </script>
-
-<!-- Add "scoped" attribute to limit CSS to this component only -->
-<style scoped>
-    .offline-card {
-        background-color: #fff5f5;
-    }
-    .fade-enter-active, .fade-leave-active {
-        transition: opacity .75s;
-    }
-    .fade-enter, .fade-leave-to /* .fade-leave-active below version 2.1.8 */ {
-        opacity: 0;
-    }
-</style>