+++ notifiers/email.go
@@ -1,17 +1,16 @@
 package notifiers
 
 import (
-	"bytes"
 	"crypto/tls"
 	"fmt"
 	"github.com/go-mail/mail"
+	"github.com/statping/emails"
 	"github.com/statping/statping/types/core"
 	"github.com/statping/statping/types/failures"
 	"github.com/statping/statping/types/notifications"
 	"github.com/statping/statping/types/notifier"
 	"github.com/statping/statping/types/services"
 	"github.com/statping/statping/utils"
-	"html/template"
 )
 
 var _ notifier.Notifier = (*emailer)(nil)
@@ -26,6 +25,10 @@ type emailer struct {
 
 func (e *emailer) Select() *notifications.Notification {
 	return e.Notification
+}
+
+func (e *emailer) Valid(values notifications.Values) error {
+	return nil
 }
 
 var email = &emailer{&notifications.Notification{
@@ -88,7 +91,7 @@ type emailOutgoing struct {
 // OnFailure will trigger failing service
 func (e *emailer) OnFailure(s services.Service, f failures.Failure) (string, error) {
 	subject := fmt.Sprintf("Service %s is Offline", s.Name)
-	tmpl := renderEmail(s, f)
+	tmpl := renderEmail(s, f, emails.Failure)
 	email := &emailOutgoing{
 		To:       e.Var2.String,
 		Subject:  subject,
@@ -101,7 +104,7 @@ func (e *emailer) OnFailure(s services.Service, f failures.Failure) (string, err
 // OnSuccess will trigger successful service
 func (e *emailer) OnSuccess(s services.Service) (string, error) {
 	subject := fmt.Sprintf("Service %s is Back Online", s.Name)
-	tmpl := renderEmail(s, failures.Failure{})
+	tmpl := renderEmail(s, failures.Failure{}, emails.Success)
 	email := &emailOutgoing{
 		To:       e.Var2.String,
 		Subject:  subject,
@@ -111,28 +114,19 @@ func (e *emailer) OnSuccess(s services.Service) (string, error) {
 	return tmpl, e.dialSend(email)
 }
 
-func renderEmail(s services.Service, f failures.Failure) string {
-	wr := bytes.NewBuffer(nil)
-	tmpl := template.New("email")
-	tmpl, err := tmpl.Parse(emailBase)
-	if err != nil {
-		log.Errorln(err)
-		return emailBase
-	}
-
+func renderEmail(s services.Service, f failures.Failure, emailData string) string {
 	data := replacer{
 		Core:    *core.App,
 		Service: s,
 		Failure: f,
 		Custom:  nil,
 	}
-
-	if err = tmpl.ExecuteTemplate(wr, "email", data); err != nil {
+	output, err := emails.Parse(emailData, data)
+	if err != nil {
 		log.Errorln(err)
-		return emailBase
+		return emailData
 	}
-
-	return wr.String()
+	return output
 }
 
 // OnTest triggers when this notifier has been saved
@@ -142,7 +136,7 @@ func (e *emailer) OnTest() (string, error) {
 	email := &emailOutgoing{
 		To:       e.Var2.String,
 		Subject:  subject,
-		Template: renderEmail(service, failures.Example()),
+		Template: renderEmail(service, failures.Example(), emailFailure),
 		From:     e.Var1.String,
 	}
 	return subject, e.dialSend(email)