+++ notifiers/notifiers.go
@@ -10,6 +10,8 @@ import (
 	"time"
 )
 
+//go:generate go run generate.go
+
 var log = utils.Log.WithField("type", "notifier")
 
 type replacer struct {
@@ -33,6 +35,7 @@ func InitNotifiers() {
 		Pushover,
 		statpingMailer,
 		Gotify,
+		AmazonSNS,
 	)
 }
 
@@ -53,10 +56,11 @@ func ReplaceTemplate(tmpl string, data replacer) string {
 
 func Add(notifs ...services.ServiceNotifier) {
 	for _, n := range notifs {
-		services.AddNotifier(n)
-		if err := n.Select().Create(); err != nil {
+		notif := n.Select()
+		if err := notif.Create(); err != nil {
 			log.Error(err)
 		}
+		services.AddNotifier(n)
 	}
 }