+++ types/configs/load.go
@@ -53,13 +53,13 @@ func LoadConfigs(cfgFile string) (*DbConfig, error) {
 	if db.Location != "" {
 		p.Set("LOCATION", db.Location)
 	}
-	if db.ApiSecret != "" {
+	if db.ApiSecret != "" && p.GetString("API_SECRET") == "" {
 		p.Set("API_SECRET", db.ApiSecret)
 	}
 	if db.Language != "" {
 		p.Set("LANGUAGE", db.Language)
 	}
-	if db.SendReports {
+	if db.AllowReports {
 		p.Set("ALLOW_REPORTS", true)
 	}
 	if db.LetsEncryptEmail != "" {
@@ -88,12 +88,11 @@ func LoadConfigs(cfgFile string) (*DbConfig, error) {
 		Location:          utils.Directory,
 		SqlFile:           p.GetString("SQL_FILE"),
 		Language:          p.GetString("LANGUAGE"),
-		SendReports:       p.GetBool("ALLOW_REPORTS"),
+		AllowReports:      p.GetBool("ALLOW_REPORTS"),
 		LetsEncryptEnable: p.GetBool("LETSENCRYPT_ENABLE"),
-	}
-	if configs.LetsEncryptEnable {
-		configs.LetsEncryptHost = p.GetString("LETSENCRYPT_HOST")
-		configs.LetsEncryptEmail = p.GetString("LETSENCRYPT_EMAIL")
+		LetsEncryptHost:   p.GetString("LETSENCRYPT_HOST"),
+		LetsEncryptEmail:  p.GetString("LETSENCRYPT_EMAIL"),
+		ApiSecret:         p.GetString("API_SECRET"),
 	}
 	log.WithFields(utils.ToFields(configs)).Debugln("read config file: " + cfgFile)